ContentProcessor and MakeHashTable works without Cheerio dependencies
