+++ src/ContentProcessor.mjs
@@ -1,21 +1,21 @@
 'use strict';
 
-import { Cheerio } from '../node_modules/cheerio/lib/cheerio.js';
+import Node from './Node.mjs';
 
 /**
  * Visits the nodes under the #content node and invoke
  * the given processor callbacks.
  *
- * @param {(root:boolean, node:{Cheerio}, isFirstPage:boolean)
+ * @param {(root:boolean, node:{Node}, isFirstPage:boolean)
  *  => void} preambleProcessor
- * @param {(root:boolean, node:{Cheerio}, partNumber:number,
+ * @param {(root:boolean, node:{Node}, partNumber:number,
  *  isFirstPage:boolean) => void} partProcessor
- * @param {(config:{object}, root:boolean, node:{Cheerio},
+ * @param {(config:{object}, root:boolean, node:{Node},
  *  thisSectLevel:number, filenamePrefix:{string},
  *  basenameFn:({string}, {number}, {number}) => {string},
  *  sectionNumber:number, isFirstPage:boolean) => void}
  *  chapterProcessor
- * @param {Cheerio} rootNode The document root node of the
+ * @param {Node} rootNode The document root node of the
  *  single html document generated by asciidoctor-chunker.
  *  The node is cloned before processing in this function
  *  to avoid the side effects.
@@ -37,18 +37,15 @@ import { Cheerio } from '../node_modules/cheerio/lib/cheerio.js';
  *  the section information.
  */
 const processContents = (
-  preambleProcessor, partProcessor, chapterProcessor, $, config, basenameMaker) => {
-  const root = $.root().clone(); // root: Cheerio<Document>
+  preambleProcessor, partProcessor, chapterProcessor, root, config, basenameMaker) => {
   let chap = 0;
   let part = 0;
   let firstPageProcessed = false;
   let isFirstPage = false;
-  root.find('#content').children().each((i, ele) => {
-    // const node = new Cheerio(ele);
-    if ($(this).hasClass('partintro'))
+  root.find('#content').children().each((node, i) => {
+    if (node.hasClass('partintro'))
       return; // ignore. this is taken care by part extraction
-    console.log("HERE");
-    if (root(this).hasClass('sect1')) {
+    if (node.hasClass('sect1')) {
       if (!firstPageProcessed && !isFirstPage) {
         isFirstPage = true;
         firstPageProcessed = true;
@@ -57,7 +54,7 @@ const processContents = (
       return chapterProcessor(config, root, node, 1, 'chap',
         basenameMaker, ++chap, isFirstPage); // recursive extraction of chapters
     }
-    if (root(this).hasClass('sect0')) {
+    if (node.hasClass('sect0')) {
       if (!firstPageProcessed && !isFirstPage) {
         isFirstPage = true;
         firstPageProcessed = true;
@@ -66,7 +63,7 @@ const processContents = (
       // part extraction
       return partProcessor(config, root, node, ++part, isFirstPage);
     }
-    if (ele.attribs.id === 'preamble') {
+    if (node.getAttr('id') === 'preamble') {
       isFirstPage = true;
       firstPageProcessed = true;
       return preambleProcessor(config, root, node, isFirstPage);
